Draft `useAgentGraph` hook
