+++ rnd/autogpt_builder/src/components/Flow.tsx
@@ -47,6 +47,7 @@ import { Control, ControlPanel } from "@/components/edit/control/ControlPanel";
 import { SaveControl } from "@/components/edit/control/SaveControl";
 import { BlocksControl } from "@/components/edit/control/BlocksControl";
 import { IconPlay, IconRedo2, IconUndo2 } from "@/components/ui/icons";
+import useAgentGraph from "@/hooks/useAgentGraph";
 
 // This is for the history, this is the minimum distance a block must move before it is logged
 // It helps to prevent spamming the history with small movements especially when pressing on a input in a block
@@ -77,59 +78,32 @@ const FlowEditor: React.FC<{
     updateEdgeData,
   } = useReactFlow<CustomNode, CustomEdge>();
   const [nodeId, setNodeId] = useState<number>(1);
-  const [availableNodes, setAvailableNodes] = useState<Block[]>([]);
-  const [savedAgent, setSavedAgent] = useState<Graph | null>(null);
-  const [agentDescription, setAgentDescription] = useState<string>("");
-  const [agentName, setAgentName] = useState<string>("");
   const [copiedNodes, setCopiedNodes] = useState<CustomNode[]>([]);
   const [copiedEdges, setCopiedEdges] = useState<CustomEdge[]>([]);
   const [isAnyModalOpen, setIsAnyModalOpen] = useState(false); // Track if any modal is open
   const [visualizeBeads, setVisualizeBeads] = useState<
     "no" | "static" | "animate"
   >("animate");
-  const [nodes, setNodes] = useState<CustomNode[]>([]);
-  const [edges, setEdges] = useState<CustomEdge[]>([]);
+  const {
+    agentName,
+    setAgentName,
+    agentDescription,
+    setAgentDescription,
+    savedAgent,
+    availableNodes,
+    getOutputType,
+    requestSave,
+    requestSaveRun,
+    nodes,
+    setNodes,
+    edges,
+    setEdges,
+  } = useAgentGraph(flowID, template);
 
-  const apiUrl = process.env.NEXT_PUBLIC_AGPT_SERVER_URL!;
-  const api = useMemo(() => new AutoGPTServerAPI(apiUrl), [apiUrl]);
   const initialPositionRef = useRef<{
     [key: string]: { x: number; y: number };
   }>({});
   const isDragging = useRef(false);
-
-  useEffect(() => {
-    api
-      .connectWebSocket()
-      .then(() => {
-        console.log("WebSocket connected");
-        api.onWebSocketMessage("execution_event", (data) => {
-          updateNodesWithExecutionData([data]);
-        });
-      })
-      .catch((error) => {
-        console.error("Failed to connect WebSocket:", error);
-      });
-
-    return () => {
-      api.disconnectWebSocket();
-    };
-  }, [api]);
-
-  useEffect(() => {
-    api
-      .getBlocks()
-      .then((blocks) => setAvailableNodes(blocks))
-      .catch();
-  }, []);
-
-  // Load existing graph
-  useEffect(() => {
-    if (!flowID || availableNodes.length == 0) return;
-
-    (template ? api.getTemplate(flowID) : api.getGraph(flowID)).then((graph) =>
-      loadGraph(graph),
-    );
-  }, [flowID, template, availableNodes]);
 
   useEffect(() => {
     const handleKeyDown = (event: KeyboardEvent) => {
@@ -192,18 +166,6 @@ const FlowEditor: React.FC<{
     delete initialPositionRef.current[node.id];
   };
 
-  const getOutputType = (id: string, handleId: string) => {
-    const node = getNode(id);
-    if (!node) return "unknown";
-
-    const outputSchema = node.data.outputSchema;
-    if (!outputSchema) return "unknown";
-
-    const outputHandle = outputSchema.properties[handleId];
-    if (!("type" in outputHandle)) return "unknown";
-    return outputHandle.type;
-  };
-
   // Function to clear status, output, and close the output info dropdown of all nodes
   // and reset data beads on edges
   const clearNodesStatusAndOutput = useCallback(() => {
@@ -243,6 +205,14 @@ const FlowEditor: React.FC<{
     },
     [deleteElements, setNodes],
   );
+
+  function formatEdgeID(conn: Link | Connection): string {
+    if ("sink_id" in conn) {
+      return `${conn.source_id}_${conn.source_name}_${conn.sink_id}_${conn.sink_name}`;
+    } else {
+      return `${conn.source}_${conn.sourceHandle}_${conn.target}_${conn.targetHandle}`;
+    }
+  }
 
   const onConnect: OnConnect = useCallback(
     (connection: Connection) => {
@@ -411,267 +381,71 @@ const FlowEditor: React.FC<{
     history.redo();
   };
 
-  function loadGraph(graph: Graph) {
-    setSavedAgent(graph);
-    setAgentName(graph.name);
-    setAgentDescription(graph.description);
+  
 
-    setNodes(() => {
-      const newNodes = graph.nodes.map((node) => {
-        const block = availableNodes.find(
-          (block) => block.id === node.block_id,
-        )!;
-        const newNode: CustomNode = {
-          id: node.id,
-          type: "custom",
-          position: {
-            x: node.metadata.position.x,
-            y: node.metadata.position.y,
-          },
-          data: {
-            block_id: block.id,
-            blockType: block.name,
-            categories: block.categories,
-            description: block.description,
-            title: `${block.name} ${node.id}`,
-            inputSchema: block.inputSchema,
-            outputSchema: block.outputSchema,
-            hardcodedValues: node.input_default,
-            connections: graph.links
-              .filter((l) => [l.source_id, l.sink_id].includes(node.id))
-              .map((link) => ({
-                edge_id: formatEdgeID(link),
-                source: link.source_id,
-                sourceHandle: link.source_name,
-                target: link.sink_id,
-                targetHandle: link.sink_name,
-              })),
-            isOutputOpen: false,
-          },
-        };
-        return newNode;
-      });
-      setEdges(
-        graph.links.map(
-          (link) =>
-          ({
-            id: formatEdgeID(link),
-            type: "custom",
-            data: {
-              edgeColor: getTypeColor(
-                getOutputType(link.source_id, link.source_name!),
-              ),
-              sourcePos: getNode(link.source_id)?.position,
-              isStatic: link.is_static,
-              beadUp: 0,
-              beadDown: 0,
-              beadData: [],
-            },
-            markerEnd: {
-              type: MarkerType.ArrowClosed,
-              strokeWidth: 2,
-              color: getTypeColor(
-                getOutputType(link.source_id, link.source_name!),
-              ),
-            },
-            source: link.source_id,
-            target: link.sink_id,
-            sourceHandle: link.source_name || undefined,
-            targetHandle: link.sink_name || undefined,
-          }),
-        ),
-      );
-      return newNodes;
-    });
-  }
+  // async function saveAgent(asTemplate: boolean = false) {
+  //   setNodes((nds) =>
+  //     nds.map((node) => ({
+  //       ...node,
+  //       data: {
+  //         ...node.data,
+  //         hardcodedValues: removeEmptyStringsAndNulls(
+  //           node.data.hardcodedValues,
+  //         ),
+  //         status: undefined,
+  //       },
+  //     })),
+  //   );
+  //   // Reset bead count
+  //   setEdges((edges) => {
+  //     return edges.map(
+  //       (edge) =>
+  //       ({
+  //         ...edge,
+  //         data: {
+  //           ...edge.data,
+  //           edgeColor: edge.data?.edgeColor!,
+  //           beadUp: 0,
+  //           beadDown: 0,
+  //           beadData: [],
+  //         },
+  //       }),
+  //     );
+  //   });
 
-  const prepareNodeInputData = (node: CustomNode) => {
-    console.log("Preparing input data for node:", node.id, node.data.blockType);
+  //   await new Promise((resolve) => setTimeout(resolve, 100));
 
-    const blockSchema = availableNodes.find(
-      (n) => n.id === node.data.block_id,
-    )?.inputSchema;
+  //   console.log("All nodes before formatting:", nodes);
+    
+  //   // SAVING
 
-    if (!blockSchema) {
-      console.error(`Schema not found for block ID: ${node.data.block_id}`);
-      return {};
-    }
+  //   // Update the node IDs in the frontend
+  //   setNodes((prev) => {
 
-    const getNestedData = (
-      schema: BlockIOSubSchema,
-      values: { [key: string]: any },
-    ): { [key: string]: any } => {
-      let inputData: { [key: string]: any } = {};
+  //     return newSavedAgent.nodes
+  //       .map((backendNode) => {
+  //         const key = `${backendNode.block_id}_${backendNode.metadata.position.x}_${backendNode.metadata.position.y}`;
+  //         const frontendNodeId = blockIdToNodeIdMap[key];
+  //         const frontendNode = prev.find((node) => node.id === frontendNodeId);
 
-      if ("properties" in schema) {
-        Object.keys(schema.properties).forEach((key) => {
-          if (values[key] !== undefined) {
-            if (
-              "properties" in schema.properties[key] ||
-              "additionalProperties" in schema.properties[key]
-            ) {
-              inputData[key] = getNestedData(
-                schema.properties[key],
-                values[key],
-              );
-            } else {
-              inputData[key] = values[key];
-            }
-          }
-        });
-      }
+  //         return frontendNode
+  //           ? {
+  //             ...frontendNode,
+  //             position: backendNode.metadata.position,
+  //             data: {
+  //               ...frontendNode.data,
+  //               backend_id: backendNode.id,
+  //             },
+  //           }
+  //           : null;
+  //       })
+  //       .filter((node) => node !== null);
+  //   });
 
-      if ("additionalProperties" in schema) {
-        inputData = { ...inputData, ...values };
-      }
+  //   await new Promise((resolve) => setTimeout(resolve, 100));
 
-      return inputData;
-    };
-
-    let inputData = getNestedData(blockSchema, node.data.hardcodedValues);
-
-    console.log(
-      `Final prepared input for ${node.data.blockType} (${node.id}):`,
-      inputData,
-    );
-    return inputData;
-  };
-
-  async function saveAgent(asTemplate: boolean = false) {
-    setNodes((nds) =>
-      nds.map((node) => ({
-        ...node,
-        data: {
-          ...node.data,
-          hardcodedValues: removeEmptyStringsAndNulls(
-            node.data.hardcodedValues,
-          ),
-          status: undefined,
-        },
-      })),
-    );
-    // Reset bead count
-    setEdges((edges) => {
-      return edges.map(
-        (edge) =>
-        ({
-          ...edge,
-          data: {
-            ...edge.data,
-            edgeColor: edge.data?.edgeColor!,
-            beadUp: 0,
-            beadDown: 0,
-            beadData: [],
-          },
-        }),
-      );
-    });
-
-    await new Promise((resolve) => setTimeout(resolve, 100));
-
-    console.log("All nodes before formatting:", nodes);
-    const blockIdToNodeIdMap: Record<string, string> = {};
-
-    const formattedNodes = nodes.map((node) => {
-      //todo kcze move outside of map!!
-      nodes.forEach((node) => {
-        const key = `${node.data.block_id}_${node.position.x}_${node.position.y}`;
-        blockIdToNodeIdMap[key] = node.id;
-      });
-      const inputDefault = prepareNodeInputData(node);
-      const inputNodes = edges
-        .filter((edge) => edge.target === node.id)
-        .map((edge) => ({
-          name: edge.targetHandle || "",
-          node_id: edge.source,
-        }));
-
-      const outputNodes = edges
-        .filter((edge) => edge.source === node.id)
-        .map((edge) => ({
-          name: edge.sourceHandle || "",
-          node_id: edge.target,
-        }));
-
-      return {
-        id: node.id,
-        block_id: node.data.block_id,
-        input_default: inputDefault,
-        input_nodes: inputNodes,
-        output_nodes: outputNodes,
-        data: {
-          ...node.data,
-          hardcodedValues: removeEmptyStringsAndNulls(
-            node.data.hardcodedValues,
-          ),
-        },
-        metadata: { position: node.position },
-      };
-    });
-
-    const links = edges.map((edge) => ({
-      source_id: edge.source,
-      sink_id: edge.target,
-      source_name: edge.sourceHandle || "",
-      sink_name: edge.targetHandle || "",
-    }));
-
-    const payload = {
-      id: savedAgent?.id!,
-      name: agentName || "Agent Name",
-      description: agentDescription || "Agent Description",
-      nodes: formattedNodes,
-      links: links,
-    };
-
-    if (savedAgent && deepEquals(payload, savedAgent)) {
-      console.debug("No need to save: Graph is the same as version on server");
-      return;
-    } else {
-      console.debug(
-        "Saving new Graph version; old vs new:",
-        savedAgent,
-        payload,
-      );
-    }
-
-    const newSavedAgent = savedAgent
-      ? await (savedAgent.is_template
-        ? api.updateTemplate(savedAgent.id, payload)
-        : api.updateGraph(savedAgent.id, payload))
-      : await (asTemplate
-        ? api.createTemplate(payload)
-        : api.createGraph(payload));
-    console.debug("Response from the API:", newSavedAgent);
-
-    // Update the node IDs in the frontend
-    setNodes((prev) => {
-      setSavedAgent(newSavedAgent);
-
-      return newSavedAgent.nodes
-        .map((backendNode) => {
-          const key = `${backendNode.block_id}_${backendNode.metadata.position.x}_${backendNode.metadata.position.y}`;
-          const frontendNodeId = blockIdToNodeIdMap[key];
-          const frontendNode = prev.find((node) => node.id === frontendNodeId);
-
-          return frontendNode
-            ? {
-              ...frontendNode,
-              position: backendNode.metadata.position,
-              data: {
-                ...frontendNode.data,
-                backend_id: backendNode.id,
-              },
-            }
-            : null;
-        })
-        .filter((node) => node !== null);
-    });
-
-    await new Promise((resolve) => setTimeout(resolve, 100));
-
-    return newSavedAgent.id;
-  }
+  //   return newSavedAgent.id;
+  // }
 
   const validateNodes = (): boolean => {
     console.log("Validating nodes");
@@ -722,131 +496,25 @@ const FlowEditor: React.FC<{
     return isValid;
   };
 
-  const runAgent = async () => {
-    try {
-      const newAgentId = await saveAgent();
-      if (!newAgentId) {
-        console.error("Error saving agent; aborting run");
-        return;
-      }
+  // const runAgent = async () => {
+  //   try {
+  //     const newAgentId = await saveAgent();
+  //     if (!newAgentId) {
+  //       console.error("Error saving agent; aborting run");
+  //       return;
+  //     }
 
-      if (!validateNodes()) {
-        console.error("Validation failed; aborting run");
-        return;
-      }
+  //     if (!validateNodes()) {
+  //       console.error("Validation failed; aborting run");
+  //       return;
+  //     }
 
-      api.subscribeToExecution(newAgentId);
-      await api.executeGraph(newAgentId);
-    } catch (error) {
-      console.error("Error running agent:", error);
-    }
-  };
-
-  function getFrontendId(backendId: string, nodes: CustomNode[]) {
-    const node = nodes.find((node) => node.data.backend_id === backendId);
-    return node?.id;
-  }
-
-  function updateEdgeBeads(
-    executionData: NodeExecutionResult[],
-    nodes: CustomNode[],
-  ) {
-    setEdges((edges) => {
-      console.log("££ Updating edges with execution data", executionData);
-      const newEdges = JSON.parse(
-        JSON.stringify(edges),
-      ) as CustomEdge[];
-
-      executionData.forEach((exec) => {
-        if (exec.status === "COMPLETED") {
-          // Produce output beads
-          for (let key in exec.output_data) {
-            const outputEdges = newEdges.filter(
-              (edge) =>
-                edge.source === getFrontendId(exec.node_id, nodes) &&
-                edge.sourceHandle === key,
-            );
-            outputEdges.forEach((edge) => {
-              edge.data!.beadUp = (edge.data!.beadUp ?? 0) + 1;
-              // For static edges beadDown is always one less than beadUp
-              // Because there's no queueing and one bead is always at the connection point
-              if (edge.data?.isStatic) {
-                edge.data!.beadDown = (edge.data!.beadUp ?? 0) - 1;
-                edge.data!.beadData! = edge.data!.beadData!.slice(0, -1);
-              }
-              //todo kcze this assumes output at key is always array with one element
-              edge.data!.beadData = [
-                exec.output_data[key][0],
-                ...edge.data!.beadData!,
-              ];
-            });
-          }
-        } else if (exec.status === "RUNNING") {
-          // Consume input beads
-          for (let key in exec.input_data) {
-            const inputEdges = newEdges.filter(
-              (edge) =>
-                edge.target === getFrontendId(exec.node_id, nodes) &&
-                edge.targetHandle === key,
-            );
-
-            inputEdges.forEach((edge) => {
-              // Skip decreasing bead count if edge doesn't match or if it's static
-              if (
-                edge.data!.beadData![edge.data!.beadData!.length - 1] !==
-                exec.input_data[key] ||
-                edge.data?.isStatic
-              ) {
-                return;
-              }
-              console.log('edge id', edge.id)
-              console.log("Consuming beadDown:", edge.data!.beadDown);
-              edge.data!.beadDown = (edge.data!.beadDown ?? 0) + 1;
-              edge.data!.beadData! = edge.data!.beadData!.slice(0, -1);
-            });
-          }
-        }
-      });
-
-      return newEdges;
-    });
-  }
-
-  const updateNodesWithExecutionData = (
-    executionData: NodeExecutionResult[],
-  ) => {
-    console.log("Updating nodes with execution data:", executionData);
-    setNodes((nodes) => {
-      if (visualizeBeads !== "no") {
-        updateEdgeBeads(executionData, nodes);
-      }
-
-      const updatedNodes = nodes.map((node) => {
-
-        const nodeExecution = executionData.find(
-          (exec) => {
-            return exec.node_id === node.data.backend_id
-          }
-        );
-
-        if (!nodeExecution || node.data.status === nodeExecution.status) {
-          return node;
-        }
-
-        return {
-          ...node,
-          data: {
-            ...node.data,
-            status: nodeExecution.status,
-            output_data: nodeExecution.output_data,
-            isOutputOpen: true,
-          },
-        };
-      });
-
-      return updatedNodes;
-    });
-  };
+  //     api.subscribeToExecution(newAgentId);
+  //     await api.executeGraph(newAgentId);
+  //   } catch (error) {
+  //     console.error("Error running agent:", error);
+  //   }
+  // };
 
   const handleKeyDown = useCallback(
     (event: KeyboardEvent) => {
@@ -939,7 +607,7 @@ const FlowEditor: React.FC<{
     {
       label: "Run",
       icon: <IconPlay />,
-      onClick: runAgent,
+      onClick: requestSaveRun,
     },
   ];
 
@@ -968,7 +636,7 @@ const FlowEditor: React.FC<{
             <BlocksControl blocks={availableNodes} addBlock={addNode} />
             <SaveControl
               agentMeta={savedAgent}
-              onSave={saveAgent}
+              onSave={(isTemplate) => requestSave(isTemplate ?? false)}
               onDescriptionChange={setAgentDescription}
               onNameChange={setAgentName}
             />
@@ -987,10 +655,4 @@ const WrappedFlowEditor: typeof FlowEditor = (props) => (
 
 export default WrappedFlowEditor;
 
-function formatEdgeID(conn: Link | Connection): string {
-  if ("sink_id" in conn) {
-    return `${conn.source_id}_${conn.source_name}_${conn.sink_id}_${conn.sink_name}`;
-  } else {
-    return `${conn.source}_${conn.sourceHandle}_${conn.target}_${conn.targetHandle}`;
-  }
-}
+
